+++ lib/models/collections/statics.dart
