+++ lib/widgets/pane_item_body.dart
